+++ app/Http/Controllers/Api/ApiV1Controller.php
@@ -70,11 +70,13 @@ class ApiV1Controller extends Controller
 			'website' 			=> 'nullable'
 		]);
 
+        $uris = implode(',', explode('\n', $request->redirect_uris));
+
         $client = Passport::client()->forceFill([
             'user_id' => null,
             'name' => e($request->client_name),
             'secret' => Str::random(40),
-            'redirect' => $request->redirect_uris,
+            'redirect' => $uris,
             'personal_access_client' => false,
             'password_client' => false,
             'revoked' => false,
@@ -828,7 +830,7 @@ class ApiV1Controller extends Controller
             ->first();
 
         if($like) {
-            $like->delete();
+            $like->forceDelete();
             $status->likes_count = $status->likes()->count();
             $status->save();
         }
@@ -1226,7 +1228,9 @@ class ApiV1Controller extends Controller
         $min = $request->input('min_id');
         $max = $request->input('max_id');
 
-        abort_if(!$since && !$min && !$max, 400);
+        if(!$since && !$min && !$max) {
+            $min = 1;
+        }
 
         $dir = $since ? '>' : ($min ? '>=' : '<');
         $id = $since ?? $min ?? $max;
@@ -1238,6 +1242,9 @@ class ApiV1Controller extends Controller
             ->limit($limit)
             ->get();
 
+        $minId = $notifications->min('id');
+        $maxId = $notifications->max('id');
+
         $resource = new Fractal\Resource\Collection(
             $notifications,
             new NotificationTransformer()
@@ -1247,7 +1254,33 @@ class ApiV1Controller extends Controller
             ->createData($resource)
             ->toArray();
 
-        return response()->json($res);
+        $baseUrl = config('app.url') . '/api/v1/notifications?';
+
+        if($minId == $maxId) {
+            $minId = null;
+        }
+
+        if($maxId) {
+            $link = '<'.$baseUrl.'max_id='.$maxId.'>; rel="next"';
+        }
+
+        if($minId) {
+            $link = '<'.$baseUrl.'min_id='.$minId.'>; rel="prev"';
+        }
+
+        if($maxId && $minId) {
+            $link = '<'.$baseUrl.'max_id='.$maxId.'>; rel="next",<'.$baseUrl.'min_id='.$minId.'>; rel="prev"';
+        }
+
+        $res = response()->json($res);
+
+        if(isset($link)) {
+            $res->withHeaders([
+                'Link' => $link,
+            ]);
+        }
+
+        return $res;
     }
 
     /**
@@ -1655,8 +1688,8 @@ class ApiV1Controller extends Controller
 
             $status = new Status;
             $status->caption = strip_tags($request->input('status'));
-            $status->scope = $request->input('visibility');
-            $status->visibility = $request->input('visibility');
+            $status->scope = $request->input('visibility', 'public');
+            $status->visibility = $request->input('visibility', 'public');
             $status->profile_id = $user->profile_id;
             $status->is_nsfw = $user->profile->cw == true ? true : $request->input('sensitive', false);
             $status->in_reply_to_id = $parent->id;
@@ -1690,8 +1723,8 @@ class ApiV1Controller extends Controller
                 abort(500, 'Invalid media ids');
             }
 
-            $status->scope = $request->input('visibility');
-            $status->visibility = $request->input('visibility');
+            $status->scope = $request->input('visibility', 'public');
+            $status->visibility = $request->input('visibility', 'public');
             $status->type = StatusController::mimeTypeCheck($mimes);
             $status->save();
         }
@@ -1756,7 +1789,9 @@ class ApiV1Controller extends Controller
         $share = Status::firstOrCreate([
             'profile_id' => $user->profile_id,
             'reblog_of_id' => $status->id,
-            'in_reply_to_profile_id' => $status->profile_id
+            'in_reply_to_profile_id' => $status->profile_id,
+            'scope' => 'public',
+            'visibility' => 'public'
         ]);
 
         if($share->wasRecentlyCreated == true) {