+++ app/Util/ActivityPub/Outbox.php
@@ -3,30 +3,49 @@
 namespace App\Util\ActivityPub;
 
 use App\Profile;
+use App\Status;
 use League\Fractal;
 use App\Http\Controllers\ProfileController;
 use App\Transformer\ActivityPub\ProfileOutbox;
+use App\Transformer\ActivityPub\Verb\CreateNote;
 
 class Outbox {
 
-	public static function get($username)
+	public static function get($profile)
 	{
         abort_if(!config('federation.activitypub.enabled'), 404);
         abort_if(!config('federation.activitypub.outbox'), 404);
 
-        $profile = Profile::whereNull('remote_url')->whereUsername($username)->firstOrFail();
         if($profile->status != null) {
             return ProfileController::accountCheck($profile);
         }
+
         if($profile->is_private) {
-            return response()->json(['error'=>'403', 'msg' => 'private profile'], 403);
+            return ['error'=>'403', 'msg' => 'private profile'];
         }
-        $timeline = $profile->statuses()->whereVisibility('public')->orderBy('created_at', 'desc')->paginate(10);
+
+        $timeline = $profile
+                    ->statuses()
+                    ->whereVisibility('public')
+                    ->orderBy('created_at', 'desc')
+                    ->take(10)
+                    ->get();
+
+        $count = Status::whereProfileId($profile->id)->count();
+
         $fractal = new Fractal\Manager();
-        $resource = new Fractal\Resource\Item($profile, new ProfileOutbox());
+        $resource = new Fractal\Resource\Collection($timeline, new CreateNote());
         $res = $fractal->createData($resource)->toArray();
 
-        return $res['data'];
+        $outbox = [
+            '@context'     => 'https://www.w3.org/ns/activitystreams',
+            '_debug'       => 'Outbox only supports latest 10 objects, pagination is not supported',
+            'id'           => $profile->permalink('/outbox'),
+            'type'         => 'OrderedCollection',
+            'totalItems'   => $count,
+            'orderedItems' => $res['data']
+        ];
+        return $outbox;
 	}
 
 }