+++ config/session.php
@@ -29,7 +29,7 @@ return [
     |
     */
 
-    'lifetime' => env('SESSION_LIFETIME', 2880),
+    'lifetime' => env('SESSION_LIFETIME', 86400),
 
     'expire_on_close' => false,