+++ resources/lang/pl/exception.php
@@ -1,0 +1,11 @@
+<?php
+
+return [
+
+	'compose' => [
+		'invalid' => [
+			'album' => 'Musi zawierać jedno zdjęcie lub film, lub wiele zdjęć.',
+		],
+	],
+
+];