+++ resources/lang/pl/profile.php
@@ -9,4 +9,7 @@ return [
   'privateProfileWarning' => 'To konto jest prywatne',
   'alreadyFollow'         => 'Już obserwujesz :username?',
   'loginToSeeProfile'     => 'aby zobaczyć zdjęcia i filmy tego użytkownika.',
+
+  'status.disabled.header' 	  => 'Profil jest niedostępny',
+  'status.disabled.body'	  => 'Przepraszamy, ten profil nie jest obecnie dostępny. Spróbuj ponownie za jakiś czas.',
 ];