+++ resources/views/site/help/hashtags.blade.php
@@ -54,7 +54,7 @@
   </div>
   <hr>
   <div class="card bg-primary border-primary" style="box-shadow: none !important;border: 3px solid #08d!important;">
-    <div class="card-header text-light font-weight-bold h4 p-4">Hashtag Tips</div>
+    <div class="card-header text-light font-weight-bold h4 p-4 bg-primary">Hashtag Tips</div>
     <div class="card-body bg-white p-3">
       <ul class="pt-3">
         <li class="lead  mb-4">You cannot add spaces or punctuation in a hashtag, or it will not work properly.</li>